updated custom node loading tree
